Various rounding type fixes
git-svn-id: https://svn.atnf.csiro.au/difx@8082 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876

+++ mpifxcorr/branches/codif/src/mk5.cpp
@@ -454,7 +454,7 @@ int Mk5DataStream::testForSync(int configindex, int buffersegment)
     if(readfromfile || readscan != 0 || readseconds != 0 || readnanoseconds != 0) //its not an error for the *first* network read
     {
       cerror << startl << "Lost Sync on segment " << buffersegment << "! Will attempt to resync. Deltatime was " << deltatime <<  "s" << endl;
-      cdebug << startl << "Corrday was " << corrday << ", corrsec was " << corrsec << ". MJD was " << mjd << ", sec was " << sec << "> Readseconds was " << bufferinfo[buffersegment].scanseconds << ". readns was " << bufferinfo[buffersegment].scanns << ", ns was " << ns << ", intclockseconds was " << intclockseconds << endl;
+      cdebug << startl << "Corrday was " << corrday << ", corrsec was " << corrsec << ". MJD was " << mjd << ", sec was " << sec << "> Readseconds was " << bufferinfo[buffersegment].scanseconds << ". readns was " << bufferinfo[buffersegment].scanns << ", ns was " << (int)ns << ", intclockseconds was " << intclockseconds << endl;
     }
     mark5stream = new_mark5_stream(
     new_mark5_stream_memory(&databuffer[buffersegment*(bufferbytes/numdatasegments)], bufferinfo[buffersegment].validbytes), new_mark5_format_generic_from_string(formatname) );
